Partially migrate special args to ints

+++ prboom2/src/hexen/p_acs.c
@@ -190,7 +190,7 @@ acsstore_t ACSStore[MAX_ACS_STORE + 1]; // +1 for termination marker
 static char EvalContext[64];
 static acs_t *ACScript;
 static unsigned int PCodeOffset;
-static byte SpecArgs[8];
+static int SpecArgs[8];
 static int ACStringCount;
 static const char **ACStrings;
 static char PrintBuffer[PRINT_BUFFER_SIZE];
